refactor(hover): контракт события отрисовывается отдельным форматтером
Sonar: связность `DescriptionFormatter` доросла до 24 классов при пороге 20, а
диагностика областей — до 21. Оба класса тянули в себя по два источника сразу.

`EventContractFormatter` берёт всё про контракт платформенного события: шапку
обработчика, параметры с типами и описаниями из синтакс-помощника, имя команды формы.
В `DescriptionFormatter` остаётся то, что человек написал в шапке-комментарии, — вместе
с этим уходят реестр типов, индекс контрактов и роли обработчиков формы.

`EventHandlerTargetRegion` вынесена из диагностики отдельной записью: имена областей
std455, выбор по роли обработчика и отступной вариант — знание про стандарт, а не про
обход дерева символов.
