+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/hover/DescriptionFormatter.java
@@ -25,16 +25,9 @@ import com.github._1c_syntax.bsl.languageserver.context.symbol.AnnotationSymbol;
 import com.github._1c_syntax.bsl.languageserver.context.symbol.MethodSymbol;
 import com.github._1c_syntax.bsl.languageserver.context.symbol.ModuleSymbol;
 import com.github._1c_syntax.bsl.languageserver.context.symbol.ParameterDefinition;
-import com.github._1c_syntax.bsl.languageserver.context.symbol.variable.VariableKind;
 import com.github._1c_syntax.bsl.languageserver.context.symbol.VariableSymbol;
-import com.github._1c_syntax.bsl.languageserver.types.index.EventContractsIndex;
-import com.github._1c_syntax.bsl.languageserver.types.model.MemberDescriptor;
-import com.github._1c_syntax.bsl.languageserver.types.model.ParameterDescriptor;
 import com.github._1c_syntax.bsl.languageserver.types.oscript.OScriptLibraryIndex;
-import com.github._1c_syntax.bsl.languageserver.configuration.Language;
 import com.github._1c_syntax.bsl.languageserver.configuration.LanguageServerConfiguration;
-import com.github._1c_syntax.bsl.languageserver.types.registry.FormHandlerRoleIndex;
-import com.github._1c_syntax.bsl.languageserver.types.registry.TypeRegistry;
 import com.github._1c_syntax.bsl.languageserver.configuration.Resources;
 import com.github._1c_syntax.bsl.parser.description.HyperlinkTypeDescription;
 import com.github._1c_syntax.bsl.parser.description.MethodDescription;
@@ -65,15 +58,10 @@ public class DescriptionFormatter {
   private static final String EXAMPLES_KEY = "examples";
   private static final String CALL_OPTIONS_KEY = "callOptions";
   private static final String DEPRECATED_FLAG_KEY = "deprecatedFlag";
-  private static final String EVENT_HANDLER_HEADER_KEY = "eventHandlerHeader";
-  private static final String COMMAND_HANDLER_HEADER_KEY = "commandHandlerHeader";
   private static final String PARAMETER_TEMPLATE = "* **%s**: %s";
 
   private final Resources resources;
   private final OScriptLibraryIndex oScriptLibraryIndex;
-  private final EventContractsIndex eventContractsIndex;
-  private final FormHandlerRoleIndex formHandlerRoleIndex;
-  private final TypeRegistry typeRegistry;
   private final LanguageServerConfiguration configuration;
 
   public void addSectionIfNotEmpty(StringJoiner markupBuilder, String newContent) {
@@ -133,231 +121,6 @@ public class DescriptionFormatter {
     }
 
     return "";
-  }
-
-  /**
-   * Секция-шапка «Обработчик события платформы: {@code <имя>}» + платформенное
-   * описание события из bsl-context. Используется hover-билдерами метода и
-   * параметра, чтобы показать, что метод/переменная — обработчик платформенного
-   * события.
-   */
-  public String getEventHandlerSection(MemberDescriptor event) {
-    return getEventHandlerSection(null, event);
-  }
-
-  /**
-   * Перегрузка с контекстным методом: к платформенному описанию события
-   * подмешивается пользовательское описание метода из шапки-комментария.
-   * Платформенное описание идёт первым, пользовательское — следом.
-   * <p>
-   * Обработчик команды формы событием не является, поэтому и называется иначе —
-   * и назван в шапке именем команды, а не процедуры: имя процедуры и так стоит
-   * в сигнатуре выше.
-   */
-  public String getEventHandlerSection(@org.jspecify.annotations.Nullable MethodSymbol method, MemberDescriptor event) {
-    var sj = new StringJoiner("\n");
-    var command = commandOf(method);
-    if (command == null) {
-      sj.add("**" + getResourceString(EVENT_HANDLER_HEADER_KEY) + ":** `" + event.name() + "`");
-    } else {
-      sj.add("**" + getResourceString(COMMAND_HANDLER_HEADER_KEY) + ":** `" + command + "`");
-    }
-    var platformDescription = event.description();
-    if (!platformDescription.isBlank()) {
-      sj.add("");
-      sj.add(platformDescription);
-    }
-    if (method != null) {
-      var userPurpose = method.getDescription()
-        .map(MethodDescription::getPurposeDescription)
-        .filter(text -> !text.isBlank())
-        .orElse("");
-      if (!userPurpose.isBlank()) {
-        sj.add("");
-        sj.add(userPurpose);
-      }
-    }
-    return sj.toString();
-  }
-
-  /**
-   * Имя команды, обработчиком которой объявлен метод.
-   *
-   * @param method метод модуля формы; {@code null} — контекста нет.
-   * @return имя команды; {@code null}, если метод — обработчик события, а не команды.
-   */
-  private @org.jspecify.annotations.Nullable String commandOf(
-    @org.jspecify.annotations.Nullable MethodSymbol method
-  ) {
-    if (method == null) {
-      return null;
-    }
-    return formHandlerRoleIndex.roleOf(method.getOwner(), method.getName())
-      .filter(handler -> handler.role() == FormHandlerRoleIndex.Role.COMMAND)
-      .map(FormHandlerRoleIndex.Handler::owner)
-      .filter(name -> !name.isBlank())
-      .orElse(null);
-  }
-
-  public String getParametersSection(MemberDescriptor eventContract) {
-    return getParametersSection(null, eventContract);
-  }
-
-  /**
-   * Перегрузка с контекстным методом: к описанию параметра из контракта
-   * подмешивается пользовательское описание из шапки-комментария метода
-   * (если оно есть). Платформенное описание идёт первым, затем пользовательское
-   * под отдельным префиксом.
-   */
-  public String getParametersSection(MethodSymbol method, MemberDescriptor eventContract) {
-    if (eventContract.signatures().isEmpty()) {
-      return "";
-    }
-    var parameters = eventContract.signatures().get(0).parameters();
-    if (parameters.isEmpty()) {
-      return "";
-    }
-    var userDescriptions = userParameterDescriptions(method);
-    var result = new StringJoiner("  \n");
-    for (var i = 0; i < parameters.size(); i++) {
-      result.add(eventParameterToString(parameters.get(i), userDescriptions, i, method));
-    }
-    var parametersSection = new StringJoiner("\n");
-    parametersSection.add("**" + getResourceString(PARAMETERS_KEY) + ":**");
-    parametersSection.add("");
-    parametersSection.add(result.toString());
-    return parametersSection.toString();
-  }
-
-  /** Имя параметра обработчика по позиции в шапке-комментарии метода (для подмешивания user-описания). */
-  private static Map<Integer, String> userParameterDescriptions(@org.jspecify.annotations.Nullable MethodSymbol method) {
-    if (method == null) {
-      return Map.of();
-    }
-    var descriptionOpt = method.getDescription();
-    if (descriptionOpt.isEmpty()) {
-      return Map.of();
-    }
-    var docParameters = descriptionOpt.get().getParameters();
-    if (docParameters.isEmpty()) {
-      return Map.of();
-    }
-    var byPosition = new HashMap<Integer, String>();
-    var methodParameters = method.getParameters();
-    for (var i = 0; i < methodParameters.size(); i++) {
-      var paramName = methodParameters.get(i).getName();
-      for (var docParam : docParameters) {
-        if (paramName.equalsIgnoreCase(docParam.name())) {
-          var purpose = docParam.types().stream()
-            .map(TypeDescription::description)
-            .filter(text -> text != null && !text.isBlank())
-            .findFirst()
-            .orElse("");
-          if (!purpose.isBlank()) {
-            byPosition.put(i, purpose);
-          }
-          break;
-        }
-      }
-    }
-    return byPosition;
-  }
-
-  private String eventParameterToString(
-    ParameterDescriptor parameter, Map<Integer, String> userDescriptions, int index,
-    @org.jspecify.annotations.Nullable MethodSymbol method
-  ) {
-    var lang = configuration.getLanguage();
-    var name = pickName(parameter, method, index, lang);
-    var types = parameter.types().refs().stream()
-      .map(ref -> typeRegistry.displayName(ref, lang))
-      .collect(Collectors.joining(" | "));
-    var line = PARAMETER_TEMPLATE.formatted(name, types);
-    var contractDescription = parameter.bilingualDescription().forLanguage(lang);
-    var userDescription = userDescriptions.getOrDefault(index, "");
-    if (!contractDescription.isBlank() && !userDescription.isBlank()) {
-      line = line + " — " + contractDescription + " / " + userDescription;
-    } else if (!contractDescription.isBlank()) {
-      line = line + " — " + contractDescription;
-    } else if (!userDescription.isBlank()) {
-      line = line + " — " + userDescription;
-    }
-    return line;
-  }
-
-  /**
-   * Имя параметра для секции «Параметры»: платформенное из контракта. Имя из кода
-   * подставляется только когда контракт его не знает — иначе разошедшиеся имена
-   * читаются как дубль: у {@code ПередЗакрытием(Отказ, СтандартнаяОбработка)} второй
-   * параметр контракта — {@code ЗавершениеРаботы}, и под именем из кода он выглядел
-   * бы вторым {@code СтандартнаяОбработка}.
-   */
-  private static String pickName(
-    ParameterDescriptor parameter, @org.jspecify.annotations.Nullable MethodSymbol method, int index,
-    Language language
-  ) {
-    var fromContract = parameter.bilingualName().forLanguage(language);
-    if (!fromContract.isBlank()) {
-      return fromContract;
-    }
-    if (method != null && index < method.getParameters().size()) {
-      var fromCode = method.getParameters().get(index).getName();
-      if (!fromCode.isBlank()) {
-        return fromCode;
-      }
-    }
-    return "";
-  }
-
-  /**
-   * Описание параметра-обработчика платформенного события из контракта
-   * (bsl-context): сопоставление <b>по позиции</b> — имена параметров обработчика
-   * задаёт пользователь, в коде они могут не совпадать с именами в контракте.
-   * При выходе за длину контракта возвращаем пусто, если последний параметр
-   * контракта не variadic.
-   */
-  public String getEventHandlerParameterDescription(VariableSymbol symbol) {
-    if (symbol.getKind() != VariableKind.PARAMETER
-      || !(symbol.getScope() instanceof MethodSymbol method)) {
-      return "";
-    }
-    var contractOpt = eventContractsIndex.getContract(method.getOwner(), method.getName());
-    if (contractOpt.isEmpty()) {
-      return "";
-    }
-    var paramIndex = indexOfParameter(method, symbol.getName());
-    if (paramIndex < 0) {
-      return "";
-    }
-    return parameterAt(contractOpt.get(), paramIndex)
-      .map(p -> p.bilingualDescription().forLanguage(configuration.getLanguage()))
-      .orElse("");
-  }
-
-  private static int indexOfParameter(MethodSymbol method, String name) {
-    var params = method.getParameters();
-    for (var i = 0; i < params.size(); i++) {
-      if (params.get(i).getName().equalsIgnoreCase(name)) {
-        return i;
-      }
-    }
-    return -1;
-  }
-
-  private static java.util.Optional<ParameterDescriptor> parameterAt(MemberDescriptor contract, int index) {
-    if (contract.signatures().isEmpty()) {
-      return java.util.Optional.empty();
-    }
-    var params = contract.signatures().get(0).parameters();
-    if (params.isEmpty()) {
-      return java.util.Optional.empty();
-    }
-    var idx = index < params.size() ? index : (params.size() - 1);
-    var p = params.get(idx);
-    if (index >= params.size() && !p.variadic()) {
-      return java.util.Optional.empty();
-    }
-    return java.util.Optional.of(p);
   }
 
   public String getReturnedValueSection(MethodSymbol methodSymbol) {
